+++ dydx3/starkex/helpers.py
@@ -6,11 +6,11 @@ from web3 import Web3
 from dydx3.constants import ASSET_RESOLUTION
 from dydx3.eth_signing.util import strip_hex_prefix
 from dydx3.starkex.constants import ORDER_FIELD_BIT_LENGTHS
-from dydx3.starkex.starkex_resources.signature import get_random_private_key
-from dydx3.starkex.starkex_resources.signature import (
+from dydx3.starkex.starkex_resources.python_signature import get_random_private_key
+from dydx3.starkex.starkex_resources.python_signature import (
     private_key_to_ec_point_on_stark_curve,
 )
-from dydx3.starkex.starkex_resources.signature import private_to_stark_key
+from dydx3.starkex.starkex_resources.python_signature import private_to_stark_key
 
 BIT_MASK_250 = (2 ** 250) - 1
 NONCE_UPPER_BOUND_EXCLUSIVE = 1 << ORDER_FIELD_BIT_LENGTHS['nonce']
